requirements.dev.txt
ipython
gitpython
